Updated arma_from_ptr copying policy
Now always copies directly from numpy, and always checks
if a conversion to f-contiguous is necessary.
When passing by const ref/ptr, a conversion is made if the numpy data
can't be used directly.
When passing by non-const ref/ptr, the numpy array won't get
through typechecking unless it's of the correct form. This should
be updated in future so it gets through typechecking but raises
an error later.